requirements.txt
anyio==3.6.1
attrs==21.4.0
bcrypt==3.2.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
ecdsa==0.18.0
fastapi==0.78.0
fastapi-login==1.8.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
